# Improve `LIKE` performance for "contains" style queries