@samuelcolvin 2024-07-28
I can the benchmarks on the exact same spec of GCP machine (`c2-standard-8`), and got slightly different results:

```
group                                              contains-performance                   master
-----                                              --------------------                   ------
ilike_utf8 scalar complex                          1.00    305.2±1.28µs        ? ?/sec    1.00    305.7±0.70µs        ? ?/sec
ilike_utf8 scalar contains                         1.11   1685.4±4.02µs        ? ?/sec    1.00   1515.0±8.44µs        ? ?/sec
ilike_utf8 scalar ends with                        1.00    233.9±0.71µs        ? ?/sec    1.00    234.3±0.31µs        ? ?/sec
ilike_utf8 scalar equals                           1.16    248.6±0.28µs        ? ?/sec    1.00    214.2±0.27µs        ? ?/sec
ilike_utf8 scalar starts with                      1.11    279.6±0.95µs        ? ?/sec    1.00    251.2±0.28µs        ? ?/sec
ilike_utf8_scalar_dyn dictionary[10] string[4])    1.00     49.1±0.22µs        ? ?/sec    1.79     88.0±0.15µs        ? ?/sec
like_utf8 scalar complex                           1.00    287.3±1.08µs        ? ?/sec    1.00    288.4±0.90µs        ? ?/sec
like_utf8 scalar contains                          1.23   458.9±11.91µs        ? ?/sec    1.00    373.6±0.36µs        ? ?/sec
like_utf8 scalar ends with                         1.00    234.4±0.21µs        ? ?/sec    1.00    234.3±0.54µs        ? ?/sec
like_utf8 scalar equals                            1.00    217.1±0.20µs        ? ?/sec    1.00    217.1±0.20µs        ? ?/sec
like_utf8 scalar starts with                       1.00    234.4±0.59µs        ? ?/sec    1.08    253.5±0.24µs        ? ?/sec
like_utf8_scalar_dyn dictionary[10] string[4])     1.00     49.0±0.07µs        ? ?/sec    1.80     88.0±0.15µs        ? ?/sec
like_utf8view scalar complex                       1.01    531.2±1.22ms        ? ?/sec    1.00    523.6±3.52ms        ? ?/sec
like_utf8view scalar contains                      1.00    134.6±0.37ms        ? ?/sec    2.97    399.7±0.46ms        ? ?/sec
like_utf8view scalar ends with                     1.01     59.0±0.21ms        ? ?/sec    1.00     58.5±0.12ms        ? ?/sec
like_utf8view scalar equals                        1.03     30.1±0.05ms        ? ?/sec    1.00     29.4±0.09ms        ? ?/sec
like_utf8view scalar starts with                   1.00     54.0±0.12ms        ? ?/sec    1.02     55.2±0.13ms        ? ?/sec
nilike_utf8 scalar complex                         1.00    304.9±0.70µs        ? ?/sec    1.00    305.2±0.59µs        ? ?/sec
nilike_utf8 scalar contains                        1.12   1684.7±5.21µs        ? ?/sec    1.00   1509.5±5.93µs        ? ?/sec
nilike_utf8 scalar ends with                       1.00    234.2±1.14µs        ? ?/sec    1.00    234.2±0.50µs        ? ?/sec
nilike_utf8 scalar equals                          1.16    248.7±0.32µs        ? ?/sec    1.00    214.2±0.43µs        ? ?/sec
nilike_utf8 scalar starts with                     1.11    279.5±0.26µs        ? ?/sec    1.00    251.3±0.47µs        ? ?/sec
nlike_utf8 scalar complex                          1.00    285.8±0.44µs        ? ?/sec    1.01    287.2±0.62µs        ? ?/sec
nlike_utf8 scalar contains                         1.24   461.9±10.51µs        ? ?/sec    1.00    373.9±0.67µs        ? ?/sec
nlike_utf8 scalar ends with                        1.00    234.2±0.20µs        ? ?/sec    1.00    234.3±0.25µs        ? ?/sec
nlike_utf8 scalar equals                           1.00    217.1±0.20µs        ? ?/sec    1.00    217.0±0.25µs        ? ?/sec
nlike_utf8 scalar starts with                      1.00    234.3±0.22µs        ? ?/sec    1.08    253.6±0.36µs        ? ?/sec
```

The only cases which should have changed in this PR are:

```
group                                              contains-performance                   master
-----                                              --------------------                   ------
like_utf8 scalar contains                          1.23   458.9±11.91µs        ? ?/sec    1.00    373.6±0.36µs        ? ?/sec
like_utf8view scalar contains                      1.00    134.6±0.37ms        ? ?/sec    2.97    399.7±0.46ms        ? ?/sec
nlike_utf8 scalar contains                         1.24   461.9±10.51µs        ? ?/sec    1.00    373.9±0.67µs        ? ?/sec
```

I'll try to work out why the first and last are marginally slower with this PR.